+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 * `mapping/star_align_v273a`: Align `.fastq` files using STAR v2.7.3a.
 
-* `mapping/star_create_reference`: Create a STAR reference index.
+* `mapping/star_build_reference`: Create a STAR reference index.
 
 * `mapping/cellranger_multi`: Align fastq files using Cell Ranger multi.